Fixing DNS client, fixing RandomIdSource and improving tests.

+++ src/main/java/com/minorityhobbies/dns/service/IdSource.java
@@ -1,5 +1,5 @@
 package com.minorityhobbies.dns.service;
 
 public interface IdSource {
-    short next();
+    int next();
 }
